requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
firebase-admin==6.5.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
streamlit-cookies-manager==0.2.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
watchdog==4.0.1
